feat(mentor): 멘토 attendance 쿼리 키 추가 및 관련 로직 개선
